Arduino Code-Prototype
This is the code Roth created for Arduino for the prototype square. The goal will be to program 64 squares in clusters (likely of 10 or 12), into 6 analog inputs. One option is to create three digit numbers as Arduino values: the first number corresponds to a cluster, and the second two numbers correspond to voltage outputs. That way, Pure Data can easily sort/split the raw data to output the wav files assigned to that cluster in that voltage bracket.

Thoughts on how to